+++ src/algoritmia/__init__.py
@@ -1,2 +1,2 @@
-TVERSION = (3, 0, 2)
+TVERSION = (3, 0, 3)
 VERSION = '.'.join([str(e) for e in TVERSION])